Update pool.js
# Security Fix
Security fix for an attack allowing to crash the nodejs process by not sending the hash result.  
Not sending the result variable lead to generating an exception.  
  
This attack can lead to and simplify a 51% attack on a coin if not patched with a coordinated attack: crashing pools and huge hashrate/corrupted daemon on the other part.  
  
This patch check that this variable exist before processing the share.  
It also prevent possible future malformed packet to crash the pool by catching every exception on the handleMessage function