Use css rules for chart colors
This adds a hidden `<a>` tag after each chart `<canvas>` with a
css-defined background-color, border-color, and width.  These are then
used by the chart javascript to set the chart's background-color,
border-color, and border-width.

The main advantage here is that it allows a site to change the chart
colors via custom.css rather than needing to edit the javascript that
generates the charts.