Added support for port-based variable donations
This adds support for allowing different donation levels to a community
donation address.  It requires adding into the blockUnlocker section of
the config file:

        "donations": {
            "enabled": true,
            "defaultDonation": 20.0,
            "address": "GDmGutfN6zbbuqqRokAFwfXakaFFFpDmS4J7aN7Bdgvaa1mLzaDi11JSADDyLnDU1oXwhgRDqFSMLJKbabPALvt2DiQiNy3"
        }

where defaultDonation is the donation amount used if the donation
amount can't be determined.  Normally donation amounts are determined
by adding a "donation" key to each port in the poolServer configuration:

            {
                "port": 3310,
                "difficulty": 5000,
                "desc": "Post description",
                "donation": 9.5
            },

which allows miners to set the donation level by choosing which port to
connect to.

This also adds the current donation level and total donations to the
telegram bot.  (website updates to report the info will follow).

This also removes my previous hack to do something similar of using the
built-in donation address--that won't work anymore since donations are
now variable per miner.