Support daemon/wallet rpc authentication; use keep-alive connections
This adds support for http authentication for the rpc connections (daemon and wallet).

It uses requests (which is added as a new dependency, thus requiring a `npm update`) rather than raw
http for json http requests.

A side-effect of this is that json http requests are now keep-alive connections rather than
establishing a new connection for every request.  (This is required for doing Digest authentication,
which is needed for monero-derived rpc authentication).