Multipool support & relay system
This adds support for two or more nodes operating as a single pool.  This required a few minor redis
changes (for things like tracking ports and active connections per-pool).

Most of the work here, however, is relaying notifications: only one host can be in control of the
bot, and e-mail may not be available on a remote VPS.  This adds support for storing notifications
on remote pools in redis, while designating one pool (i.e. the master) to pick up those messages and
relay them via telegram or e-mail.