Slush mining updates
This fixes some aspects of slush rewards:

- when slush rewards were enabled actual shares were only estimated; this
commit splits up the scores and shares values so that accurate shares can still
be reported.

- when an orphan was found during unlocking, the shares (i.e. score) of miners
was re-added to the current score.  This is just plain wrong under slush mining
(it's kind of weird under prop mining too) since it's transplating shares 2
hours into the future.  This isn't great under prop, but can be a disaster
under slush (particularly when the orphaned block took exceptionally long since
the orphaned scores will dominate the current block scores).  This one could
probably be fixed by rescaling the scores of the orphaned block by `exp((τ - t)
/ λ)`, where `τ` is the orphaned block's start time and `t` is the new block's
start time, but since that is pretty much guaranteed to be a number very close
to 0 (since this is happening when unlocking, i.e. 2 hours later) it's not
going to do much, so I just skipped replanting the shares entirely under score
payments.

- Slush scores can easily exceed the limits of an int; store them as floats
instead so that we don't have to worry about such an overflow.

- The recorded share data had a serious bug where it could end up recording
infinite for a worker's share: the code requested the last block time, but
didn't wait for the response before using it, which would result in a division
by 0, thus inserting infinity for a user's score.  This is bad (every reward is
going to come out as 0 or NaN).

- To deal with the above, the slush calculation is now done using a lua command
(i.e. a redis `EVAL`) to do the last block lookup in the same transaction as
the score insertion.  This crucially also avoids a race condition: if a block
is discovered between the time of the last block time lookup and the score
insertion the score would be far too high and dominate the block (at least
until the new block reached the same age as the old block).

- The above also eliminates the need for last-block-time polling for score
entirely: it isn't needed at all now since the lua eval gets the value,
calculates the exponential score, and increments the share atomically.

- Scores/shares are now included in the API and reported on the worker stats
page.

- Added a blurb to the Payments page describing the payment mechanism.

- remove blockTime integral calculation approximations from javascript; they
were fairly inaccurate (particularly in the face of pool hashrate
fluctuations), and more importantly the actual shares are available now.