Add Load More; update and fix share percent
Instead of just putting a hard cut-off at 20, this adds a Load More
button to load an addition 20 each time it is clicked.

Enhanced the share percentage output to show the user's share before
and after the pool fee is removed.

Fixed a bug with the percentage not being rewritten if needed.

Used an event for the JS popup rather than an inline javascript string.