Add ability to use first vout for block reward
Loki, in particular, no longer uses a constant fee: most blocks don't
include the governance reward, but then there is an occassional very
large reward with includes a large governance reward.  The proper
solution here is to use the first vout to extract the miner reward,
which this patch adds support for.